refactor: extract Checked In handler from handleStateTransitions
Move the Checked In branch into effects/checkedInEffects.ts. Handler
stamps checkedInAt / checkedInBy and persists the XState snapshot
to Firestore so statusFromXState resolves correctly on reload. The
rest of check-in processing is triggered by /api/checkin-processing
from the client.

HandlerContext gains an actor field so handlers can call
getPersistedSnapshot() without the dispatcher synthesizing the
snapshot itself.

No behavior change. tsc clean, 44 tests pass.
